requirements.txt
google-cloud-storage
google-cloud-bigquery
pandas
apache-airflow-providers-google
google-cloud-pubsub
apache-beam[gcp]
faker
